power: Convert to fully binderized 1.1 implementation
power hal: Add power HAL API 1.1 impl for Wahoo

Add a full binderized implementation for Power hal

Many subsystems (e.g.wifi) could be living on an
independent power island (sourced from VBatt directly)
and might even have their own dedicated XTAL to source
their clocks. Since these SOCs are capable of
autonomously operating (while the platform is in one
of the sleep states), they are still drawing power
from the VBatt. Hence it is critical to understand
the  (SOC) level low power statistics as well when
the battery level changes and be able to find any
correlation in event of unexpected battery drain.

This commit adds the support of the Power Hal 1.1
to wahoo based devices(that includes Muskie/walleye).
This includes the new api for wlan specific power stats

Bug: 29339696
Test: Manual
Change-Id: Iee4e38f2d9ced31f8b6a333b535fa1d9a302ec26
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>

Re-introduce of POWER HAL API 1.1 impl for Wahoo

Power HAL 1.1 support for wahoo based devices was initially introduced in CL
ag/2098359
However, this caused a regression in application startup times due
to a bug in passing parameters for power hints on application launch
Hence, that CL was reverted in CL ag/2270791

This commit brings back the support of the Power Hal 1.1
to wahoo based devices. This includes the changes of the original CL
as will as a fix for the app startup time regression

The fix is similar to that in ag/1767023 (done for power HAL 1.0 default
implementation)
where a NULL is passed to the powerHint function when the passed data is
Zero
(instead of passing a pointer to the data). This enable the App Launch
power hints to work properly

The commit has been tested not to cause that regression

Bug: 62040325
Test: Performance Tests along with other tests

Change-Id: I29ce38b2de92c2b9ad878b0076288b689695b8a0
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>